一緒に演奏されている曲を分割
